added delete permissions

+++ helpdesk/prevence/users.php
@@ -10,27 +10,31 @@ if (isset($_SESSION['admin_id'])) {
    header('location:index.php');
 }
 if (isset($_POST['delete_user'])) {
-   if(isset($_POST['user_id'])){
-      $user_id = $_POST['user_id'];
-      $delete_query = mysqli_query($conn, "DELETE FROM `tickets` WHERE userId = $user_id");
-      if($delete_query){
-         $delete_query = mysqli_query($conn, "DELETE FROM `users` WHERE userId = $user_id");
-         $message[] = "User deleted successfully";
-      } else {
-         $message[] = "Error deleting user";
+   if ($_SESSION['department'] == 'All') {
+      if (isset($_POST['user_id'])) {
+         $user_id = $_POST['user_id'];
+         $delete_query = mysqli_query($conn, "DELETE FROM `tickets` WHERE userId = $user_id");
+         if ($delete_query) {
+            $delete_query = mysqli_query($conn, "DELETE FROM `users` WHERE userId = $user_id");
+            $message[] = "User deleted successfully";
+         } else {
+            $message[] = "Error deleting user";
+         }
+      } else if (isset($_POST['admin_id'])) {
+         $admin_id = $_POST['admin_id'];
+         $delete_query = mysqli_query($conn, "DELETE FROM `admins` WHERE adminId = $admin_id");
+         if ($delete_query) {
+            $message[] = "Admin deleted successfully";
+         } else {
+            $message[] = "Error deleting admin";
+         }
+
+
+      }
+   } else {
+      $message[] = "You don't have permission to delete users";
    }
- }
-   else if(isset($_POST['admin_id'])){
-      $admin_id = $_POST['admin_id'];
-      $delete_query = mysqli_query($conn, "DELETE FROM `admins` WHERE adminId = $admin_id");
-      if($delete_query){
-         $message[] = "Admin deleted successfully";
-      } else {
-         $message[] = "Error deleting admin";
-   }
-   
-  
- }
+
 }
 
 
@@ -43,7 +47,7 @@ if (isset($_POST['delete_user'])) {
    <meta charset="UTF-8">
    <meta http-equiv="X-UA-Compatible" content="IE=edge">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-   <title>Users</title>
+   <title>účty</title>
 
    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
 
@@ -67,55 +71,57 @@ if (isset($_POST['delete_user'])) {
                <div class="inputBox">
                   <select name="users" class="select" required>
                      <option value="all">--- Choose a user type ---</option>
-                     <option <?php if($_POST['users'] == 'frontend') echo 'selected'?> value="frontend">Frontend</option>
-                     <option <?php if($_POST['users'] == 'backend') echo 'selected'?> value="backend">Backend</option>
-                  </select>
+                     <option <?php if ($_POST['users'] == 'frontend')
+                        echo 'selected' ?> value="frontend">Frontend</option>
+                        <option <?php if ($_POST['users'] == 'backend')
+                        echo 'selected' ?> value="backend">Backend</option>
+                     </select>
+                  </div>
+                  <div class="inputBox">
+                     <button type="submit" name="filter" class="btn">Filter</button>
+                  </div><br>
                </div>
-               <div class="inputBox">
-                  <button type="submit" name="filter" class="btn">Filter</button>
-               </div><br>
-            </div>
 
-         <div class="box-container">
-            <?php
-             $frontendUsers = returnAllFrontendUsers($conn);
-             $backendUsers = returnAllBackendUsers($conn);
-            if (isset($_POST['filter']) && $_POST['users'] == 'frontend' || ($_POST['users'] == 'all' || !isset($_POST['filter']))) {
-               foreach ($frontendUsers as $user) {
-                  echo '<div class="box">
-                  <div class="breaking"><p> ID : <span>'.$user['userId'].'</span> </p></div>
-                  <div class="breaking"><p> Name : <span>'.$user['userName'].'</span> </p></div>
-                  <div class="breaking"><p> Surname : <span>'.$user['userSurname'].'</span> </p></div>
-                  <div class="breaking"><p> Email : <span>'.$user['userEmail'].'</span> </p></div>
+               <div class="box-container">
+                  <?php
+                     $frontendUsers = returnAllFrontendUsers($conn);
+                     $backendUsers = returnAllBackendUsers($conn);
+                     if (isset($_POST['filter']) && $_POST['users'] == 'frontend' || ($_POST['users'] == 'all' || !isset($_POST['filter']))) {
+                        foreach ($frontendUsers as $user) {
+                           echo '<div class="box">
+                  <div class="breaking"><p> ID : <span>' . $user['userId'] . '</span> </p></div>
+                  <div class="breaking"><p> Name : <span>' . $user['userName'] . '</span> </p></div>
+                  <div class="breaking"><p> Surname : <span>' . $user['userSurname'] . '</span> </p></div>
+                  <div class="breaking"><p> Email : <span>' . $user['userEmail'] . '</span> </p></div>
                   <form method="POST">
                   <input type="hidden" name="user_id" value="' . $user['userId'] . '"><br>
                   <button type="submit" name="delete_user" class="delete-btn">Delete</button>
                   </form>
                   </div>';
-               }
-            }
-            if (isset($_POST['filter']) && $_POST['users'] == 'backend' || ($_POST['users'] == 'all' || !isset($_POST['filter']))) {
-               foreach ($backendUsers as $user) {
-                  echo '<div class="box">
-                  <div class="breaking"><p> ID : <span>'.$user['adminId'].'</span> </p></div>
-                  <div class="breaking"><p> Name : <span>'.$user['adminName'].'</span> </p></div>
-                  <div class="breaking"><p> Surname : <span>'.$user['adminSurname'].'</span> </p></div>
-                  <div class="breaking"><p> Email : <span>'.$user['adminEmail'].'</span> </p></div>
-                  <div class="breaking"><p> Department : <span>'.returnDepartmentName($conn, $user['departmentId'])['departmentName'].'</span> </p></div>
+                        }
+                     }
+                     if (isset($_POST['filter']) && $_POST['users'] == 'backend' || ($_POST['users'] == 'all' || !isset($_POST['filter']))) {
+                        foreach ($backendUsers as $user) {
+                           echo '<div class="box">
+                  <div class="breaking"><p> ID : <span>' . $user['adminId'] . '</span> </p></div>
+                  <div class="breaking"><p> Name : <span>' . $user['adminName'] . '</span> </p></div>
+                  <div class="breaking"><p> Surname : <span>' . $user['adminSurname'] . '</span> </p></div>
+                  <div class="breaking"><p> Email : <span>' . $user['adminEmail'] . '</span> </p></div>
+                  <div class="breaking"><p> Department : <span>' . returnDepartmentName($conn, $user['departmentId'])['departmentName'] . '</span> </p></div>
                   <form method="POST">
                   <input type="hidden" name="admin_id" value="' . $user['adminId'] . '"> <br>
                   <button type="submit" name="delete_user" class="delete-btn">Delete</button>
                   </form>
                   </div>';
-               }
-            }
-            //if there are no users
-            if (count($frontendUsers) == 0 && count($backendUsers) == 0){
-               echo '<p class="empty">No users</p>';
-            }
+                        }
+                     }
+                     //if there are no users
+                     if (count($frontendUsers) == 0 && count($backendUsers) == 0) {
+                        echo '<p class="empty">No users</p>';
+                     }
 
-            ?>
-         </div>
+                     ?>
+            </div>
 
       </section>
    </section>
@@ -124,5 +130,3 @@ if (isset($_POST['delete_user'])) {
 </body>
 
 </html>
-
-
